refactor: document and rename canvasUtils
