+++ packages/editor-ui/src/composables/useNodeConnections.test.ts
@@ -3,7 +3,7 @@ import { NodeConnectionType } from 'n8n-workflow';
 import { useNodeConnections } from '@/composables/useNodeConnections';
 import type { CanvasNodeData } from '@/types';
 import { CanvasConnectionMode } from '@/types';
-import { createCanvasConnectionHandleString } from '@/utils/canvasUtilsV2';
+import { createCanvasConnectionHandleString } from '@/utils/canvasUtils';
 
 describe('useNodeConnections', () => {
 	const defaultConnections = {
